[#176339507] Reworked tests with local stub of PM
